requirements.txt
Django==3.2
djangorestframework==3.12.2
matplotlib==3.3.4
